apps/worker/requirements-dev.txt
-r requirements.txt
fakeredis==2.28.1
pytest==8.3.4
